requirements.txt
ella>=3.0.5,<4
django-jsonfield
django-tastypie==0.9.14
django-object-permissions
